# deploy/openapi.json
{
  "components": {
    "schemas": {
...
      "EnrollRequest": {
        "additionalProperties": false,
        "properties": {
          "age": {
            "minimum": 0.0,
            "title": "Age",
            "type": "integer"
          },
          "athlete_id": {
            "minLength": 1,
            "title": "Athlete Id",
            "type": "string"
          },
          "bodyweight_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyweight Kg"
          },
          "experience": {
            "title": "Experience",
            "type": "string"
          },
          "sex": {
            "title": "Sex",
            "type": "string"
          }
        },
        "required": [
          "athlete_id",
          "sex",
          "age",
          "experience"
        ],
        "title": "EnrollRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionStartRequest": {
        "description": "POST /sessions \u2014 no athlete-supplied composition parameters (composition is server-owned\nand load-free; the client cannot influence it). Idempotent on client_request_id.",
        "properties": {
          "client_request_id": {
            "minLength": 1,
            "title": "Client Request Id",
            "type": "string"
          }
        },
        "required": [
          "client_request_id"
        ],
        "title": "SessionStartRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Hush v1 \u2014 advisory training intelligence. Hush recommends; the athlete decides.",
    "title": "Hush v1 API",
    "version": "v1",
    "x-hush-schema-version": 11
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/athletes": {
      "post": {
        "description": "Enroll an athlete and mint their bearer token. The athlete is created via the existing\nonboarding seed (ES-008 v2 / Option D); the token's hash is stored, the plaintext returned\nonce for the secure provisioning channel (BB-25).",
        "operationId": "enroll_internal_athletes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enroll",
        "tags": [
          "internal"
        ]
      }
    },
...
    "/sessions": {
      "get": {
        "operationId": "sessions_list_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sessions List",
        "tags": [
          "read"
        ]
      },
      "post": {
        "operationId": "start_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "sessions"
        ]
      }
    },
...
  }
}